start to modulation of seperate components of project + fixes and improvements
