+++ README.md
@@ -1,0 +1,1 @@
+npm install react-select
