+++ src/App.js
@@ -1,0 +1,11 @@
+import Nav from './components/Nav';
+function App() {
+  return (
+    <div className='App'>
+      <Nav />
+      
+    </div>
+  )
+}
+
+export default App;
