+++ src/components/InputBox.js
@@ -1,0 +1,12 @@
+const InputBox = () => {
+    return (
+        <form className="Input_body">
+            <input type="text" name="input" />
+
+            <input type="submit" value="submit" />
+        </form>
+    );
+
+};
+
+export default InputBox;
