+++ src/components/styles/nav.css
@@ -1,0 +1,58 @@
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100..900&display=swap');  
+
+/* background of navigation bar */
+.main-nav {
+    background-color: #D9D9D9;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    margin: auto;
+    width: 70%;
+    height: 80px;
+    align-items: center;
+}
+
+.nav-format {
+    display: flex;
+    flex-direction: row;
+    list-style-type: none;
+    justify-content: flex-start;
+    height: 100%;
+    width: 90%;
+    gap: 100px;
+    align-items: center;
+}
+
+.nav-title {
+    font-family: 'Inter';
+    font-style: normal;
+    font-weight: 800;
+    font-size: 16px;
+    text-transform: uppercase;
+    color: #528B8E;
+    flex-grow: 2;
+
+}
+
+.nav-type {
+    font-family: 'Inter';
+    font-style: normal;
+    font-weight: 500;
+    font-size: 16px;
+    color: #579397;
+
+}
+
+
+.nav-about {
+    font-family: 'Inter';
+    font-weight: 700;
+    font-size: 18px;
+    color: #528B8E;
+    line-height: 40px;
+    text-align: center;
+
+    background-color: #CBCBCB;
+    height: 40px;
+    width: 40px;
+    backdrop-filter: blur(2px);
+    color: #528B8E;
+}
